--- src/main/resources/cadastro.fxml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="477.0" prefWidth="268.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.controller.PessoaController">
-   <children>
-      <AnchorPane prefHeight="481.0" prefWidth="268.0">
-         <children>
-            <Label layoutX="24.0" layoutY="39.0" text="Código *" />
-            <TextField fx:id="idUsuario" layoutX="21.0" layoutY="55.0" prefHeight="25.0" prefWidth="73.0" />
-            <Label fx:id="mensagemIdUsuario" layoutX="23.0" layoutY="80.0" prefHeight="17.0" prefWidth="225.0" textAlignment="CENTER" textFill="#f40909" />
-            <Label layoutX="27.0" layoutY="108.0" text="Nome *" />
-            <TextField fx:id="nome" layoutX="22.0" layoutY="124.0" prefHeight="25.0" prefWidth="225.0" />
-            <Label fx:id="mensagemNome" layoutX="26.0" layoutY="150.0" prefHeight="17.0" prefWidth="217.0" textAlignment="CENTER" textFill="#f40909" />
-            <Label layoutX="25.0" layoutY="178.0" text="CPF *" />
-            <TextField fx:id="cpf" layoutX="22.0" layoutY="195.0" prefHeight="25.0" prefWidth="225.0" />
-            <Label fx:id="mensagemCpf" layoutX="26.0" layoutY="221.0" prefHeight="17.0" prefWidth="217.0" textAlignment="CENTER" textFill="#f40909" />
-            <Label layoutX="28.0" layoutY="247.0" text="Endereço" />
-            <TextField fx:id="endereco" layoutX="24.0" layoutY="264.0" prefHeight="25.0" prefWidth="225.0" />
-            <Label layoutX="30.0" layoutY="316.0" text="Telefone" />
-            <TextField fx:id="telefone" layoutX="26.0" layoutY="333.0" />
-            <Button fx:id="salvar" layoutX="95.0" layoutY="393.0" mnemonicParsing="false" onAction="#executarSalvar" text="Salvar" />
-         </children>
-      </AnchorPane>
-   </children>
-</VBox>
