--- src/main/resources/main.fxml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="477.0" prefWidth="268.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.controller.UsuarioController">
-   <children>
-      <AnchorPane prefHeight="481.0" prefWidth="268.0">
-         <children>
-            <Label layoutX="35.0" layoutY="123.0" text="Usuário *" />
-            <TextField fx:id="nomeUsuario" layoutX="35.0" layoutY="142.0" prefHeight="25.0" prefWidth="184.0" />
-            <Label fx:id="mensagemUsuario" layoutX="38.0" layoutY="168.0" prefHeight="17.0" prefWidth="184.0" text="" textAlignment="CENTER" textFill="#f10404" />
-            <Label layoutX="35.0" layoutY="193.0" text="Senha *" />
-            <TextField fx:id="senha" layoutX="35.0" layoutY="210.0" prefHeight="25.0" prefWidth="184.0" />
-            <Label fx:id="mensagemSenha" layoutX="37.0" layoutY="236.0" prefHeight="17.0" prefWidth="184.0" text="" textAlignment="CENTER" textFill="#f10404" />
-            <Button fx:id="login" layoutX="101.0" layoutY="268.0" mnemonicParsing="false" onAction="#executarLoginUsuario" text="Login" />
-         </children>
-      </AnchorPane>
-   </children>
-</VBox>
