+++ src/main/resources/usuario.fxml
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?import javafx.scene.text.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="516.0" prefWidth="295.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.controller.UsuarioController">
+   <children>
+      <TabPane prefHeight="516.0" prefWidth="295.0" tabClosingPolicy="UNAVAILABLE">
+         <tabs>
+            <Tab fx:id="entrar" text="Login">
+               <content>
+                  <AnchorPane prefHeight="200.0" prefWidth="200.0">
+                     <children>
+                        <Label layoutX="46.0" layoutY="123.0" text="Usuário *" />
+                        <TextField fx:id="nomeUsuario" layoutX="45.0" layoutY="140.0" prefHeight="25.0" prefWidth="206.0" />
+                        <Label fx:id="mensagemUsuario" layoutX="45.0" layoutY="166.0" prefHeight="17.0" prefWidth="206.0" textFill="#e70808">
+                           <font>
+                              <Font size="8.0" />
+                           </font>
+                        </Label>
+                        <Label layoutX="46.0" layoutY="193.0" text="Senha *" />
+                        <TextField fx:id="senha" layoutX="45.0" layoutY="209.0" prefHeight="25.0" prefWidth="205.0" />
+                        <Label fx:id="mensagemSenha" layoutX="45.0" layoutY="236.0" prefHeight="17.0" prefWidth="205.0" textFill="#e70808">
+                           <font>
+                              <Font size="8.0" />
+                           </font>
+                        </Label>
+                        <Button fx:id="botaoLogin" layoutX="122.0" layoutY="263.0" mnemonicParsing="false" onAction="#executarBotaoLogin" text="Login" />
+                     </children>
+                  </AnchorPane>
+               </content>
+            </Tab>
+            <Tab text="Cadastro">
+               <content>
+                  <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
+                     <children>
+                        <Label layoutX="27.0" layoutY="42.0" text="Código *" />
+                        <TextField fx:id="idUsuario" layoutX="23.0" layoutY="60.0" prefHeight="25.0" prefWidth="75.0" />
+                        <Label fx:id="mensagemCodigo" layoutX="24.0" layoutY="85.0" prefHeight="17.0" prefWidth="246.0" textFill="#e70808">
+                           <font>
+                              <Font size="8.0" />
+                           </font>
+                        </Label>
+                        <Label layoutX="27.0" layoutY="110.0" text="Nome *" />
+                        <TextField fx:id="nome" layoutX="23.0" layoutY="126.0" prefHeight="25.0" prefWidth="250.0" />
+                        <Label fx:id="mensagemNome" layoutX="24.0" layoutY="152.0" prefHeight="17.0" prefWidth="248.0" textFill="#e70808">
+                           <font>
+                              <Font size="8.0" />
+                           </font>
+                        </Label>
+                        <Label layoutX="25.0" layoutY="178.0" text="CPF *" />
+                        <TextField fx:id="cpf" layoutX="23.0" layoutY="194.0" prefHeight="25.0" prefWidth="248.0" />
+                        <Label layoutX="26.0" layoutY="244.0" text="Endereço" />
+                        <TextField fx:id="endereco" layoutX="23.0" layoutY="261.0" prefHeight="22.0" prefWidth="248.0" />
+                        <Label layoutX="28.0" layoutY="314.0" text="Telefone" />
+                        <TextField fx:id="telefone" layoutX="23.0" layoutY="330.0" />
+                        <Button fx:id="salvar" layoutX="121.0" layoutY="388.0" mnemonicParsing="false" onAction="#executarBotaoSalvar" text="Salvar" />
+                        <Label fx:id="mensagemCpf" layoutX="24.0" layoutY="219.0" prefHeight="17.0" prefWidth="247.0" textFill="#e70808">
+                           <font>
+                              <Font size="8.0" />
+                           </font>
+                        </Label>
+                     </children>
+                  </AnchorPane>
+               </content>
+            </Tab>
+         </tabs>
+      </TabPane>
+   </children>
+</VBox>
